refactor: remove mixin, move demo logic
